Refactor edge difference iterator to work from table collections as well as from tree sequences.
